FlexMediaTrait::deleteMediaFile() should check if folder exists

+++ system/src/Grav/Framework/Flex/Traits/FlexMediaTrait.php
@@ -160,6 +160,10 @@ trait FlexMediaTrait
 
         $fileParts  = pathinfo($basename);
 
+        if (!file_exists($targetPath)) {
+            return;
+        }
+
         if (file_exists($targetFile)) {
 
             $result = unlink($targetFile);
